Refactor: 자동 객체 관리를 통해 명시적인 삭제 작업 제거 및 CollabPostRequestDTO 필드타입 수정 (#19)

+++ src/main/java/com/partnerd/web/dto/collabDTO/request/CollabPostRequestDTO.java
@@ -24,7 +24,7 @@ public class CollabPostRequestDTO {
         private int eventMode;
         private String description;
         private List<ContactMethodDTO> contactMethod;
-        private List<Integer> categoryIds;
+        private List<Long> categoryIds;
     }
